Output stacktrace on crush
